Move IntrusiveObjectMixin implementation to source file
